+++ src/timber_design/ghpython/components/CT_BeamDecompose/code.py
@@ -1,4 +1,3 @@
-# r: timber_design>=0.1.0
 """Extracts main geometric characteristics of a Beam."""
 
 # flake8: noqa