+++ src/timber_design/ghpython/components/CT_ShowTopologyTypes/code.py
@@ -1,16 +1,9 @@
-# r: timber_design>=0.1.0
 """Shows the names of the connection topology types."""
 
 # flake8: noqa
-import re
 import Grasshopper
 import System
-from compas_rhino.conversions import point_to_rhino
-from compas_timber.connections import ConnectionSolver
-from compas_timber.connections import GenericJoint
 from compas_timber.connections import get_clusters_from_joint_candidates
-from compas_timber.connections import JointTopology
-from compas_timber.utils import intersection_line_line_param
 from timber_design.ghpython.ghcomponent_helpers import item_input_valid_cpython